Working search bar
One thing worth paying attention to is that I wrapped another layer 'shownItems' outside of 'itemData.shoppingList', in order to show the search results while searching and go back to the full list after searching. So now in the html, the shown list is looping through 'shownItems' instead of 'itemData.shoppingList' directly.